% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VariableBinningS3.R
\name{plot.VariableBinningS3}
\alias{plot.VariableBinningS3}
\title{plot.VariableBinningS3}
\usage{
\method{plot}{VariableBinningS3}(obj)
}
\arguments{
\item{obj}{An object of type \code{VariableBinningS3}.}
}
\description{
Implements the \code{plot} method for \code{VariableBinningS3}.
}
\examples{
plot()
}
