% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.VariableBinningS3.R
\name{predict.VariableBinningS3}
\alias{predict.VariableBinningS3}
\title{predict.VariableBinningS3}
\usage{
\method{predict}{VariableBinningS3}(obj, variableVector)
}
\arguments{
\item{obj}{An object of type \code{VariableBinningS3}.}
}
\description{
Implements the \code{predict} method for \code{VariableBinningS3}.
}
\examples{
plot()
}
