% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableSummary.R
\name{variableSummary}
\alias{variableSummary}
\title{Provides a summary of the independent variable versus the response.}
\usage{
variableSummary(data, target, variable)
}
\arguments{
\item{data}{The dataframe of interest.}

\item{target}{A string value containing the name of the binary target variable.}

\item{variable}{A string value containing the name of the independent variable.}
}
\description{
Provides a summary of the independent variable versus the response.
}
\examples{
variableSummary()
}
