% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightOfEvidence.R
\name{weightOfEvidence}
\alias{weightOfEvidence}
\title{Wight of Evidence Function}
\usage{
weightOfEvidence(data, target, variable, cuts)
}
\arguments{
\item{data}{The dataframe of interest.}

\item{target}{A string value containing the name of the binary target variable.}

\item{variable}{A string value containing the name of the independent variable.}

\item{cuts}{Either a vector of continuous cutpoints, or a list of categorical groupings by which to group the independent variable.}
}
\description{
This function takes a data frame , along with a target variable name, independent variable name, and user defined
groupings in order to produce Weight of Evidence bandings and transformations.
It returns an S3 object which implements plot and predict methods.
}
\examples{
cat_function()
}
